Giphy is now set up.
Added Giphy API stuff. For now the search query is hard coded as
“Sunny,” and we don’t want the query to be like that. We need to figure
out how to have OWM send the weather description to Giphy to query. To
do that, I’m pretty sure we need to figure out how to turn the OWM
description into a variable that can be appended to the end of the
Giphy API web call.