The site now has a working weather description variable
After days of fiddling, I finally figured out how to alert the
description of the weather, showing that I have the weather description
functioning. I missed the fact that in the JSON file from OWM, the
Weather.weather stuff was inside of an array. This is why I was able to
access every other element on the page but anything inside of this
object.