It's alive!
At long last, I have the two APIs talking to each other. I figured out
how to use a callback to get the Giphy API to wait for the user to type
in their search.

Additionally, I have made it so that every time the user searches, it
gives them a random gif. I also added a basic Bootstrap theme.

Looking forward, issues include styling the gif (due to the fact the
gif is not in the HTML), search suggestions and the fact that the user
has to reload the page every time they want to do a new search.